+++ aws-sdk-core/lib/aws-sdk-core/dynamodbstreams.rb
@@ -1,0 +1,4 @@
+Aws.add_service(:DynamoDBStreams, {
+  api: "#{Aws::API_DIR}/streams.dynamodb/2012-08-10/api-2.json",
+  docs: "#{Aws::API_DIR}/streams.dynamodb/2012-08-10/docs-2.json",
+})