+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Aws::DynamoDBStreams - Added support for Amazon DynamoDB Streams.
+
 * Feature - Aws::DynamoDB - Added support for consistent reads with the
   Scan API.